requirements.txt
python-telegram-bot
langchain-groq
langchain
python-dotenv
